requirements.txt
fastapi
uvicorn
httpx
PyGithub
anthropic
tree-sitter==0.25.0
tree-sitter-python
tree-sitter-javascript
python-dotenv
pydantic
sqlalchemy[asyncio]
aiosqlite
